+++ cas-injector/src/main/java/com/oreki/cas_injector/core/dto/AggregatedHolding.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -11,18 +12,18 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AggregatedHolding {
     private String schemeName;      // 1
-    private double units;           // 2
-    private double currentValue;    // 3
-    private double investedAmount;  // 4
-    private double ltcgValue;       // 5 (The sellable cash)
-    private double ltcgAmount;      // 6 (The profit in LTCG)
-    private double stcgValue;       // 7 (The locked cash)
-    private double stcgAmount;      // 8 (The profit in STCG)
-    private int daysToNextLtcg;     // 9 (Fixed from 0)
+    private BigDecimal units;           // 2
+    private BigDecimal currentValue;    // 3
+    private BigDecimal investedAmount;  // 4
+    private BigDecimal ltcgValue;       // 5
+    private BigDecimal ltcgAmount;      // 6
+    private BigDecimal stcgValue;       // 7
+    private BigDecimal stcgAmount;      // 8
+    private int daysToNextLtcg;     // 9
     private int oldestAgeDays;      // 10
     private String assetCategory;   // 11
     private String status;          // 12
-    private String isin;            // 13 (Added for your Strategy check)
-    private Double stcgTaxEstimate; // 14 (Added for Phase 1E)
-    private double nav;             // 15 (Added to preserve precision instead of dividing currentValue/units)
+    private String isin;            // 13
+    private BigDecimal stcgTaxEstimate; // 14
+    private BigDecimal nav;             // 15
 }
