+++ cas-injector/src/main/java/com/oreki/cas_injector/stocks/StockHoldingDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -15,21 +16,21 @@ public class StockHoldingDTO {
     private String companyName;
     private String exchange;
     private String sector;
-    private double quantity;
-    private double avgCostPerShare;
-    private double currentPrice;
-    private double currentValue;
-    private double investedAmount;
-    private double unrealisedPnl;
-    private double unrealisedPnlPct;
-    private double realisedGainFy;
-    private double unrealisedLtcg;
-    private double unrealisedStcg;
-    private double ltcgTaxEstimate;
-    private double stcgTaxEstimate;
+    private BigDecimal quantity;
+    private BigDecimal avgCostPerShare;
+    private BigDecimal currentPrice;
+    private BigDecimal currentValue;
+    private BigDecimal investedAmount;
+    private BigDecimal unrealisedPnl;
+    private BigDecimal unrealisedPnlPct;
+    private BigDecimal realisedGainFy;
+    private BigDecimal unrealisedLtcg;
+    private BigDecimal unrealisedStcg;
+    private BigDecimal ltcgTaxEstimate;
+    private BigDecimal stcgTaxEstimate;
     private int daysToNextLtcg;
-    private double xirr;
-    private double dayChangePct;
+    private BigDecimal xirr;
+    private BigDecimal dayChangePct;
     private String action;
-    private double convictionScore;
+    private BigDecimal convictionScore;
 }
